Restructure kratos UI
