+++ src/components/Authentication/Kratos/KratosProps.ts
@@ -1,0 +1,11 @@
+import { UiNode } from '@ory/kratos-client';
+
+export interface KratosProps {
+  node: UiNode;
+}
+
+export interface KratosInputExtraProps {
+  autoComplete?: string;
+  autoCapitalize?: string;
+  autoCorrect?: string;
+}
